==> code/02_python_example.py <==
# -*- coding: utf-8 -*-
"""
Created on Sun Jan 18 12:40:20 2015

@author: Bryan
"""

# instantiating the list of strings
test_list = ['a', 'b', 'c', 'd']

# print the first three items in the list
print test_list[0:3]

# printing the entire list of strings
for item in test_list:
    print item

#appending to the list of strings
test_list.append('e')
